Restructuring the project so that it can still build with angular-cli but also be packaged with ng-packagr.
