+++ ng-package.json
@@ -1,0 +1,14 @@
+{
+  "$schema": "./node_modules/ng-packagr/ng-package.schema.json",
+  "whitelistedNonPeerDependencies": ["."],
+  "lib": {
+    "entryFile": "public_api.ts",
+    "umdModuleIds": {
+      "lodash" : "_",
+      "d3-interpolate": "d3",
+      "toxiclibsjs": "toxicLibsJS",
+      "twemoji": "twemoji",
+      "gsap": "gsap"
+    }
+  }
+}
