+++ package.json
@@ -13,24 +13,29 @@
     "build:prod": "npm run setup && ng build --env=prod -prod -aot=false -extract-css=false -vc=false",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "packagr": "ng-packagr -p ng-package.json"
   },
-  "dependencies": {
-    "@angular/animations": "^5.0.2",
-    "@angular/cdk": "^5.0.0-rc0",
+  "peerDependencies": {
+    "@angular/animations": "^5.2.0",
+    "@angular/cdk": "^5.0.4",
     "@angular/common": "^5.0.2",
     "@angular/compiler": "^5.0.2",
     "@angular/core": "^5.0.2",
     "@angular/elements": "github:angular/elements-builds#labs/elements",
     "@angular/forms": "^5.0.1",
     "@angular/http": "^5.0.2",
-    "@angular/material": "^5.0.0-rc0",
+    "@angular/material": "^5.0.4",
     "@angular/platform-browser": "^5.0.1",
     "@angular/platform-browser-dynamic": "^5.0.1",
-    "@angular/router": "^5.0.1",
+    "@angular/router": "^5.0.1"
+  },
+  "dependencies": {
     "@conversationai/perspectiveapi-simple-server": "^0.1.4",
+    "@types/d3-interpolate": "^1.1.5",
+    "@types/gapi": "^0.0.31",
     "@types/gsap": "^1.19.2",
-    "@types/lodash": "^4.14.85",
+    "@types/lodash": "^4.14.108",
     "@webcomponents/custom-elements": "^1.0.6",
     "concurrently": "^3.4.0",
     "core-js": "^2.4.1",
@@ -39,7 +44,6 @@
     "gsap": "^1.19.1",
     "hammerjs": "^2.0.8",
     "jasmine": "~2.8.0",
-    "js-sha256": "^0.5.0",
     "lodash": "^4.17.4",
     "ngx-color-picker": "^4.3.0",
     "node-emoji": "^1.8.1",
@@ -49,10 +53,20 @@
     "zone.js": "^0.8.5"
   },
   "devDependencies": {
+    "@angular/animations": "^5.2.0",
+    "@angular/cdk": "^5.0.4",
     "@angular/cli": "~1.5.5",
+    "@angular/common": "^5.0.2",
+    "@angular/compiler": "^5.0.2",
     "@angular/compiler-cli": "^5.0.1",
-    "@types/d3-interpolate": "^1.1.5",
-    "@types/gapi": "^0.0.31",
+    "@angular/core": "^5.0.2",
+    "@angular/elements": "github:angular/elements-builds#labs/elements",
+    "@angular/forms": "^5.0.1",
+    "@angular/http": "^5.0.2",
+    "@angular/material": "^5.0.4",
+    "@angular/platform-browser": "^5.0.1",
+    "@angular/platform-browser-dynamic": "^5.0.1",
+    "@angular/router": "^5.0.1",
     "@types/jasmine": "~2.8.2",
     "@types/node": "~9.6.5",
     "jasmine-core": "~2.8.0",
@@ -65,6 +79,7 @@
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-mocha": "^1.3.0",
     "mocha": "^4.0.1",
+    "ng-packagr": "^2.4.2",
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
