+++ src/app/modules/convai-checker/convai-checker.component.css
