+++ src/app/modules/convai-checker/perspective-status.component.css
