+++ src/app/modules/convai-checker/perspectiveapi-types.ts
