+++ src/main_with_elements.ts
@@ -3,10 +3,10 @@ import 'gsap';
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import {ElementDemoModule} from './app/element-demo.module';
+import {ElementDemoModule} from './app/modules/convai-checker/element-demo.module';
 import { environment } from './environments/environment';
 import { registerAsCustomElements } from '@angular/elements';
-import {ConvaiChecker} from './app/convai-checker.component';
+import {ConvaiChecker} from './app/modules/convai-checker/convai-checker.component';
 
 if (environment.production) {
   enableProdMode();
