+++ tsconfig.json
@@ -4,7 +4,7 @@
     "outDir": "./build/out-tsc",
     "baseUrl": "src",
     "sourceMap": true,
-    "declaration": false,
+    "declaration": true,
     "module": "es2015",
     "moduleResolution": "node",
     "emitDecoratorMetadata": true,
@@ -17,5 +17,12 @@
       "es2016",
       "dom"
     ]
+  },
+  "angularCompilerOptions": {
+    "strictMetadataEmit": true,
+    "skipTemplateCodegen": true,
+    "flatModuleOutFile": "./build/perspectiveapi-authorship-demo.js",
+    "flatModuleId": "perspectiveapi-authorship-demo-lib",
+    "annotateForClosureCompiler": true
   }
 }
